vng: Use uint64 for Data Lengths (#5320)

+++ vng/primitive.go
@@ -141,9 +141,9 @@ func (p *PrimitiveEncoder) Metadata(off uint64) (uint64, Metadata) {
 		}
 	}
 	loc := Segment{
-		Offset:            int64(off),
-		Length:            int32(len(p.out)),
-		MemLength:         int32(p.bytesLen),
+		Offset:            off,
+		Length:            uint64(len(p.out)),
+		MemLength:         p.bytesLen,
 		CompressionFormat: p.format,
 	}
 	off += uint64(len(p.out))
